[*] Fix all compiler warnings

+++ src/debug.c
@@ -1,3 +1,24 @@
+/*
+ *      This file is part of the SmokeOS project.
+ *  Copyright (C) 2015  <Fabien Bavent>
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ *   - - - - - - - - - - - - - - -
+ *
+ *      Routines to help kernel debugging or monitoring.
+ */
 #include <smkos/kapi.h>
 #include <smkos/kstruct/fs.h>
 #include <smkos/kstruct/map.h>
@@ -85,7 +106,7 @@ void kstacktrace(size_t MaxFrames)
 /**
     Display hexadeciaml data of a memory area.
  */
-void kdump (void *ptr, size_t lg)
+void kdump (void *ptr, int lg)
 {
   int i;
